Update VT online deadline
The state technically allows online registration
until election day, but says it may not be 
included on the paper list at the polling place,
in which case the registrant may need to register
again in person before voting. They request but
do not strictly seem to require that you do online
registration by the Friday before. This is a bit 
splitting hairs.